+++ deobsf/v26_1/src/forge/java/org/adde0109/pcf/v26_1/forge/Config.java
@@ -1,4 +1,4 @@
-package org.adde0109.pcf.v16_5.forge;
+package org.adde0109.pcf.v26_1.forge;
 
 import static org.adde0109.pcf.PCF.CONFIG_FILE_NAME;
