+++ modern/v14_4/build.gradle.kts
@@ -20,8 +20,7 @@ unimined.minecraft(forge) {
 
 dependencies {
     forgeCompileOnly(libs.mixin)
-    evaluationDependsOn(":modern:v16_5")
-    forgeCompileOnly(srcSetAsDep(":modern:v16_5", "forge"))
+    forgeCompileOnly(srcSetAsDep(":deobsf:v26_1", "forge"))
     forgeCompileOnly(project(":common"))
 }
