+++ modern/v19_2/build.gradle.kts
@@ -19,9 +19,6 @@ unimined.minecraft(forge) {
 }
 
 dependencies {
-    evaluationDependsOn(":modern:v20_4")
-    forgeCompileOnly(srcSetAsDep(":modern:v16_5", "forge"))
-    forgeCompileOnly(srcSetAsDep(":modern:v20_4", "forge"))
     forgeCompileOnly(project(":common"))
 }
