+++ modern/v20_4/build.gradle.kts
@@ -19,7 +19,7 @@ unimined.minecraft(forge) {
 }
 
 dependencies {
-    forgeCompileOnly(srcSetAsDep(":modern:v16_5", "forge"))
+    forgeCompileOnly(srcSetAsDep(":deobsf:v26_1", "forge"))
     forgeCompileOnly(project(":common"))
 }
