+++ modern/v20_4/src/forge/java/org/adde0109/pcf/v20_4/forge/ReloadConfigInit.java
@@ -13,7 +13,7 @@ import net.minecraftforge.fml.javafmlmod.FMLModContainer;
 
 import org.adde0109.pcf.PCF;
 import org.adde0109.pcf.PCFInitializer;
-import org.adde0109.pcf.v16_5.forge.Config;
+import org.adde0109.pcf.v26_1.forge.Config;
 
 @AConstraint(
         platform = Platform.FORGE,
@@ -21,13 +21,13 @@ import org.adde0109.pcf.v16_5.forge.Config;
 public final class ReloadConfigInit implements PCFInitializer {
     @Override
     public void onInit() {
-        IEventBus eventBus =
+        final IEventBus eventBus =
                 MetaAPI.instance()
                         .meta()
                         .<FMLModContainer>mod(PCF.MOD_ID)
                         .map(Wrapped::unwrap)
                         .map(FMLModContainer::getEventBus)
                         .orElseThrow();
-        eventBus.addListener((ModConfigEvent.Reloading event) -> Config.reload());
+        eventBus.addListener((ModConfigEvent.Reloading _) -> Config.reload());
     }
 }
